+++ Packages/XCharts/Runtime/Component/Axis/Axis.cs
@@ -99,6 +99,7 @@ namespace XCharts.Runtime
         [SerializeField] protected bool m_Inverse = false;
         [SerializeField] private bool m_Clockwise = true;
         [SerializeField] private bool m_InsertDataToHead;
+        [SerializeField][Since("v3.11.0")] private float m_MinCategorySpacing = 0;
         [SerializeField] protected List<Sprite> m_Icons = new List<Sprite>();
         [SerializeField] protected List<string> m_Data = new List<string>();
         [SerializeField] protected AxisLine m_AxisLine = AxisLine.defaultAxisLine;
@@ -405,6 +406,15 @@ namespace XCharts.Runtime
         {
             get { return m_InsertDataToHead; }
             set { if (PropertyUtil.SetStruct(ref m_InsertDataToHead, value)) SetAllDirty(); }
+        }
+        /// <summary>
+        /// The minimum spacing between categories.
+        /// ||类目之间的最小间距。
+        /// </summary>
+        public float minCategorySpacing
+        {
+            get { return m_MinCategorySpacing; }
+            set { if (PropertyUtil.SetStruct(ref m_MinCategorySpacing, value)) SetAllDirty(); }
         }
 
         public override bool vertsDirty
@@ -934,6 +944,44 @@ namespace XCharts.Runtime
                     axisLength :
                     (float)(Math.Abs(context.minValue) * (axisLength / (Math.Abs(context.minValue) + Math.Abs(context.maxValue))))
                 );
+        }
+
+        public Vector3 GetCategoryPosition(int categoryIndex, int dataCount = 0)
+        {
+            if (dataCount <= 0)
+            {
+                dataCount = data.Count;
+            }
+            if (IsCategory() && dataCount > 0)
+            {
+                Vector3 pos;
+                if (boundaryGap)
+                {
+                    var each = context.length / dataCount;
+                    pos = context.start + context.dire * (each * (categoryIndex + 0.5f));
+                }
+                else
+                {
+                    var each = context.length / (dataCount - 1);
+                    pos = context.start + context.dire * (each * categoryIndex);
+                }
+                if (axisLabel.distance != 0)
+                {
+                    if (this is YAxis)
+                    {
+                        pos.x = GetLabelObjectPosition(0).x;
+                    }
+                    else
+                    {
+                        pos.y = GetLabelObjectPosition(0).y;
+                    }
+                }
+                return pos;
+            }
+            else
+            {
+                return Vector3.zero;
+            }
         }
     }
 }