+++ Packages/XCharts/Runtime/Component/Tooltip/TooltipHandler.cs
@@ -65,6 +65,7 @@ namespace XCharts.Runtime
 
                 m_LabelRoot = ChartHelper.AddObject("label", tooltip.gameObject.transform, chart.chartMinAnchor,
                     chart.chartMaxAnchor, chart.chartPivot, chart.chartSizeDelta);
+                m_LabelRoot.transform.SetSiblingIndex(0);
                 ChartHelper.HideAllObject(m_LabelRoot);
                 m_IndicatorLabels.Clear();
                 foreach (var com in chart.components)
@@ -169,6 +170,11 @@ namespace XCharts.Runtime
                     }
                 }
             }
+            if (!anyTrigger && m_ContainerSeries == null)
+            {
+                m_ContainerSeries = ListPool<Serie>.Get();
+                UpdatePointerContainerAndSeriesAndTooltip(tooltip, ref m_ContainerSeries);
+            }
             if (m_ContainerSeries != null)
             {
                 if (!SetSerieTooltip(tooltip, m_ContainerSeries))
@@ -180,7 +186,7 @@ namespace XCharts.Runtime
             }
             if (!m_ShowTooltip || !anyTrigger)
             {
-                if (tooltip.context.type == Tooltip.Type.Corss && m_PointerContainer != null && m_PointerContainer.IsPointerEnter())
+                if (tooltip.context.type == Tooltip.Type.Cross && m_PointerContainer != null && m_PointerContainer.IsPointerEnter())
                 {
                     m_ShowTooltip = true;
                     tooltip.SetActive(true);
@@ -205,7 +211,7 @@ namespace XCharts.Runtime
             if (tooltip.context.type == Tooltip.Type.None) return;
             if (m_PointerContainer != null)
             {
-                if (tooltip.context.type == Tooltip.Type.Corss)
+                if (tooltip.context.type == Tooltip.Type.Cross)
                 {
                     if (m_PointerContainer is GridCoord)
                     {
@@ -512,7 +518,7 @@ namespace XCharts.Runtime
             {
                 axis.context.axisTooltipValue = serie.GetSerieData(serie.context.pointerItemDataIndex).GetData(dimension);
             }
-            else if (component.type == Tooltip.Type.Corss)
+            else if (component.type == Tooltip.Type.Cross)
             {
                 axis.context.axisTooltipValue = axis.context.pointerValue;
             }
@@ -692,11 +698,11 @@ namespace XCharts.Runtime
                     float splitWidth = AxisHelper.GetDataWidth(xAxis, grid.context.width, dataCount, dataZoom);
                     switch (tooltip.context.type)
                     {
-                        case Tooltip.Type.Corss:
+                        case Tooltip.Type.Cross:
                         case Tooltip.Type.Line:
                             float pX = grid.context.x;
                             pX += xAxis.IsCategory() ?
-                                (float)(xAxis.context.axisTooltipValue * splitWidth + (xAxis.boundaryGap ? splitWidth / 2 : 0)) :
+                                (float)(xAxis.context.pointerValue * splitWidth + (xAxis.boundaryGap ? splitWidth / 2 : 0)) :
                                 xAxis.GetDistance(xAxis.context.axisTooltipValue, grid.context.width);
                             if (pX < grid.context.x)
                                 break;
@@ -704,7 +710,7 @@ namespace XCharts.Runtime
                             Vector2 ep = new Vector2(pX, grid.context.y + grid.context.height);
                             var lineColor = TooltipHelper.GetLineColor(tooltip, chart.theme.tooltip.lineColor);
                             ChartDrawer.DrawLineStyle(vh, lineType, lineWidth, sp, ep, lineColor);
-                            if (tooltip.context.type == Tooltip.Type.Corss)
+                            if (tooltip.context.type == Tooltip.Type.Cross)
                             {
                                 sp = new Vector2(grid.context.x, chart.pointerPos.y);
                                 ep = new Vector2(grid.context.x + grid.context.width, chart.pointerPos.y);
@@ -758,7 +764,7 @@ namespace XCharts.Runtime
                     float splitWidth = AxisHelper.GetDataWidth(yAxis, grid.context.height, dataCount, dataZoom);
                     switch (tooltip.context.type)
                     {
-                        case Tooltip.Type.Corss:
+                        case Tooltip.Type.Cross:
                         case Tooltip.Type.Line:
                             float pY = (float)(grid.context.y + yAxis.context.pointerValue * splitWidth +
                                 (yAxis.boundaryGap ? splitWidth / 2 : 0));
@@ -768,7 +774,7 @@ namespace XCharts.Runtime
                             Vector2 ep = new Vector2(grid.context.x + grid.context.width, pY);
                             var lineColor = TooltipHelper.GetLineColor(tooltip, chart.theme.tooltip.lineColor);
                             ChartDrawer.DrawLineStyle(vh, lineType, lineWidth, sp, ep, lineColor);
-                            if (tooltip.context.type == Tooltip.Type.Corss)
+                            if (tooltip.context.type == Tooltip.Type.Cross)
                             {
                                 sp = new Vector2(chart.pointerPos.x, grid.context.y);
                                 ep = new Vector2(chart.pointerPos.x, grid.context.y + grid.context.height);
@@ -813,7 +819,7 @@ namespace XCharts.Runtime
 
             switch (tooltip.context.type)
             {
-                case Tooltip.Type.Corss:
+                case Tooltip.Type.Cross:
                     ChartDrawer.DrawLineStyle(vh, lineType, lineWidth, sp, ep, lineColor);
                     var dist = Vector2.Distance(chart.pointerPos, cenPos);
                     if (dist > radius) dist = radius;