+++ Packages/XCharts/Documentation~/zh/changelog.md
@@ -73,6 +73,13 @@ slug: /changelog
 
 ## master
 
+* (2024.06.11) 修复`Axis`的`IndicatorLabel`可能会遮挡住`Tooltip`的问题
+* (2024.06.11) 修复`Tooltip`在`Cross`时`Axis`的`IndicatorLabel`可能不显示的问题 (#315)
+* (2024.06.10) 调整`Tooltip`的`Corss`重命名为`Cross`
+* (2024.06.09) 增加`Axis`的`minCategorySpacing`设置类目轴默认的最小类目间距
+* (2024.06.09) 修复`Tooltip`的`Cross`在`Axis`是类目轴并且开启`DataZoom`的情况下指示位置不准确的问题
+* (2024.06.06) 修复`Serie`在`Clone`时动画异常问题 (#320)
+* (2024.06.04) 修复`Serie`的`state`在代码动态设置时不刷新的问题
 * (2024.05.29) 调整`XCharts`在`Hierarchy`视图下的右键菜单到`UI/XCharts`下
 * (2024.05.29) 增加`3D`坐标系对类目轴的支持
 * (2024.05.19) 优化`Editor`下的编辑性能